requirements.txt
# Web Framework
Flask>=2.3.0

# AI & ML
google-generativeai>=0.3.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0

# Vector Search for RAG
faiss-cpu>=1.7.4
xlrd
